Update suggested plugins to new API
