+++ setup.cfg
@@ -51,9 +51,9 @@ libtbx.precommit =
 workflows.services =
     Schlockmeister = zocalo.service.schlockmeister:Schlockmeister
 zocalo.configuration.plugins =
-    graylog = zocalo.configuration.graylog:GraylogPlugin
-    jmx = zocalo.configuration.jmx
-    stomp = zocalo.configuration.stomp
+    graylog = zocalo.configuration.plugin_graylog:Graylog
+    jmx = zocalo.configuration.plugin_jmx:JMX
+    stomp = zocalo.configuration.plugin_stomp:Stomp
 zocalo.wrappers =
     dummy = zocalo.wrapper:DummyWrapper
