+++ src/zocalo/configuration/plugin_graylog.py
@@ -1,0 +1,13 @@
+from marshmallow import fields, validate
+
+from zocalo.configuration import PluginSchema
+
+
+class Graylog(PluginSchema):
+    protocol = fields.Str(validate=validate.OneOf(["UDP", "TCP"]), required=True)
+    host = fields.Str(required=True)
+    port = fields.Int(required=True)
+
+    @staticmethod
+    def activate(configuration):
+        raise NotImplementedError("Make a thing happen!")
